Renamed constant.py to settings.py
